Split testcase from integration test file
